Move `ClarityVersion` and `TransactionPayload`
