+++ clarity-types/src/lib.rs
@@ -27,11 +27,13 @@ pub use stacks_common::{
 pub mod errors;
 pub mod representations;
 pub mod types;
+pub mod version;
 
 pub use errors::{ClarityTypeError, IncomparableError};
 pub use representations::{ClarityName, ContractName};
 use stacks_common::types::StacksEpochId;
 pub use types::Value;
+pub use version::ClarityVersion;
 
 /// Max call stack depth for Epoch 3.4+.
 const MAX_CALL_STACK_DEPTH: u64 = 128;
